+++ src/java/rmi/dgc/VMID.java
@@ -67,7 +67,7 @@ public final class VMID implements java.io.Serializable {
      * conditions: a) the conditions for uniqueness for objects of
      * the class <code>java.rmi.server.UID</code> are satisfied, and b) an
      * address can be obtained for this host that is unique and constant
-     * for the lifetime of this object.
+     * for the lifetime of this object.  <p>
      */
     public VMID() {
         addr = randomBytes;
@@ -123,7 +123,7 @@ public final class VMID implements java.io.Serializable {
         StringBuffer result = new StringBuffer();
         if (addr != null)
             for (int i = 0; i < addr.length; ++ i) {
-                int x = addr[i] & 0xFF;
+                int x = (int) (addr[i] & 0xFF);
                 result.append((x < 0x10 ? "0" : "") +
                               Integer.toString(x, 16));
             }
